tkdjango/requirements.txt
Django==4.2.13
djangorestframework==3.16.0
django-cors-headers==4.7.0
django-extensions==4.1
pillow==9.5.0
python-decouple==3.8
pytz==2025.2
sqlparse==0.5.3
tzdata==2025.2
pymysql
zeep